Introduce `copy-invite-token` make command
Since `make get-invite-token` typically is followed by copying the token
to paste into the web interface, I automated that.

This required updating `issue_invite_for_the_frontend` to respect the
styling that is implemented in `django.core.management.base.BaseCommand`
- `ERROR` probably isn't conceptually the right designation for this
value, but it results in (as far as my not-a-visual-designer eyes can
tell) the same styling.

---

Also corrects a typo in a comment: `eqaul -> equal`