Update wording when creating an integration (#1572)
Hello everyone 👋 

This is my first PR to the OnCall repo so please let me know if I'm not
doing something right. I'm also not sure if the items in the Checklist
are for me to complete or someone else. If those are for me, please let
me know and I'll be happy to do them (with some guidance as a
first-timer)!

# What this PR does
This PR is a really small suggestion to the wording in this file. I
believe it will help improve the way this line of text is read by users.

## Which issue(s) this PR fixes

## Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated

---------

Co-authored-by: Joey Orlando <joseph.t.orlando@gmail.com>
Co-authored-by: Joey Orlando <joey.orlando@grafana.com>