Fix helm hook for migration job (#1789)
# What this PR does
This PR fixes the migration job when using helm hooks.
## Which issue(s) this PR fixes
Currently when enabling the useHooks option in the helm chart there is a
deadlock as the hook is defined as "post-upgrade, post-install". In this
case the database migration will only be executed when all other pods
are ready. This will not happen as the pods are waiting for the database
migration. So we need to use "pre-install, preupgrade" instead of
"post-install,post-upgrade"

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [ ] Documentation added (or `pr:no public docs` PR label added if not
required)
- [ x] `CHANGELOG.md` updated (or `pr:no changelog` PR label added if
not required)

Co-authored-by: Ildar Iskhakov <Ildar.iskhakov@grafana.com>