Update all links to use `docs/reference` style
The `docs/reference` shortcode supports contextual destinations and version inference.

`<ONCALL VERSION>` is inferred to match the version of the documentation set.
For example, the inferred version for the page /docs/grafana/oncall/latest/get-started/ is "latest".
It can also be overriden using front matter.
Given the same page, but with the additional front matter `oncall_version: next`, the variable is substituted with "next" rather than "latest".

This is used by the Grafana Cloud documentation to always link to the "next" version of OnCall documentation which is expected to match the code deployed in Grafana Cloud more closely than "latest".
If this is not correct, please let me know so I can make the appropriate changes in the website.

Contextual destinations are achieved using repeated labels in the shortcode inner text.
The format is [<LABEL>]: "<PAGE PATH PREFIX> -> <HUGO REFERENCE>".

- _`<LABEL>`_ matches the reference style link label used in the rest of the text.
- _`<PAGE PATH PREFIX>`_ is matched against the page during the production build.
  If the match is successful, the destination that is used is _`<HUGO REFERENCE>`_.
  The first matching prefix is used, not the longest matching prefix.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>