adjust frontmatter weight and add menu title (#2484)
The doc squad is restructuring the Grafana Cloud TOC to align with UI
navigation. This PR makes minor changes to the source _index.md file
frontmatter in support of these changes.

Changed weight to 500 to manipulate order or doc sets and added
menuTitle field to remove "Grafana" from TOC title for consistency.