Update all make docs CI to match (#2538)
My previous PR only updated the CI that ran on PRs, I forgot about the
CD for releases.

Fixes https://github.com/grafana/oncall/actions/runs/5547818896.

You can see that this will resolve the problem because it is what allows
the PR CI to pass. I just forgot to include it in the release CD.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>