Remove docker-compose warnings (#2574)
Remove docker-compose warnings when running hobby deployment without
Prometheus exporter enabled.
Otherwise, I'm getting the following warnings when running any
docker-compose command:

```bash
docker-compose up -d
WARN[0000] The "FEATURE_PROMETHEUS_EXPORTER_ENABLED" variable is not set. Defaulting to a blank string.
WARN[0000] The "PROMETHEUS_EXPORTER_SECRET" variable is not set. Defaulting to a blank string.
WARN[0000] The "FEATURE_PROMETHEUS_EXPORTER_ENABLED" variable is not set. Defaulting to a blank string.
WARN[0000] The "PROMETHEUS_EXPORTER_SECRET" variable is not set. Defaulting to a blank string.
WARN[0000] The "FEATURE_PROMETHEUS_EXPORTER_ENABLED" variable is not set. Defaulting to a blank string.
WARN[0000] The "PROMETHEUS_EXPORTER_SECRET" variable is not set. Defaulting to a blank string.
WARN[0000] The "FEATURE_PROMETHEUS_EXPORTER_ENABLED" variable is not set. Defaulting to a blank string.
WARN[0000] The "PROMETHEUS_EXPORTER_SECRET" variable is not set. Defaulting to a blank string.
[+] Building 0.0s (0/0)
[+] Running 6/6
 ✔ Network oncall-oss_default                  Created                                                           0.0s
 ✔ Container oncall-oss-grafana-1              Started                                                           0.3s
 ✔ Container oncall-oss-redis-1                Healthy                                                           6.5s
 ✔ Container oncall-oss-oncall_db_migration-1  Exited                                                            7.5s
 ✔ Container oncall-oss-engine-1               Started                                                           7.6s
 ✔ Container oncall-oss-celery-1               Started                                                           7.6s
```