Tweaks for integrations (#2869)
# What this PR does

This PR adds a few tweaks to the integrations page

- prevent calling `/counters` on the integration page, instead call it
with `counters/[id]`
- call to `/counters` will be done only once when we load all the
integrations and prevent calling it again once we switch pages. Prior to
this change every page resulted in a call to `/counters`, and due to the
fact the request took on average `3-5s` it was being stalled by the
browser on subsequent calls.
- Call to `/connected_contact_points` is now conditioned by the
integration type
- No more page flicking when going back from integration page to the
list of integrations
- No more page flickering when switching pages in table view if there's
more requests ongoing
- Prevent discarding filters when going back from Integration view to
the table view
- Fixed routes not moving up/down
- Fixed routes flickering when adding/removing a route

These changes should result in slightly loader times because `/counters`
was time-consuming request, and less UX issues.

## Which issue(s) this PR fixes

- described at https://github.com/grafana/oncall/issues/2843 but
incompletely

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [ ] Documentation added (or `pr:no public docs` PR label added if not
required)
- [ ] `CHANGELOG.md` updated (or `pr:no changelog` PR label added if not
required)